+++ src/rules/declaration-colon-newline-after/index.js
@@ -12,7 +12,7 @@ export const ruleName = "declaration-colon-newline-after"
 export const messages = ruleMessages(ruleName, {
   expectedAfter: () => "Expected newline after \":\"",
   rejectedAfter: () => "Unexpected whitespace after \":\"",
-  expectedAfterMultiLine: () => "Unexpected whitespace after \":\" with a multi-line value",
+  expectedAfterMultiLine: () => "Expected newline after \":\" with a multi-line value",
 })
 
 export default function (expectation) {
@@ -33,15 +33,19 @@ export default function (expectation) {
 
       // Get the raw prop, and only the prop
       const endOfPropIndex = declarationValueIndexOffset(decl) + decl.raw("between").length - 1
-      const declString = decl.toString().slice(0, endOfPropIndex)
 
-      for (let i = 0, l = declString.length; i < l; i++) {
-        if (declString[i] !== ":") { continue }
-        const indexToCheck = (declString.substr(declString[i], 3) === "/*")
-          ? declString.indexOf("*/", i) + 1
+      // The extra characters tacked onto the end ensure that there is a character to check
+      // after the colon. Otherwise, with `background:pink` the character after the
+      const propPlusColon = decl.toString().slice(0, endOfPropIndex) + "xxx"
+
+      for (let i = 0, l = propPlusColon.length; i < l; i++) {
+        if (propPlusColon[i] !== ":") { continue }
+        const indexToCheck = (propPlusColon.substr(propPlusColon[i], 3) === "/*")
+          ? propPlusColon.indexOf("*/", i) + 1
           : i
+
         checker.afterOneOnly({
-          source: declString,
+          source: propPlusColon,
           index: indexToCheck,
           lineCheckStr: decl.value,
           err: m => {