+++ src/rules/declaration-colon-space-after/index.js
@@ -44,12 +44,15 @@ export function declarationColonSpaceChecker({ locationChecker, root, result, ch
 
     // Get the raw prop, and only the prop
     const endOfPropIndex = declarationValueIndexOffset(decl) + decl.raw("between").length - 1
-    const declString = decl.toString().slice(0, endOfPropIndex)
 
-    for (let i = 0, l = declString.length; i < l; i++) {
-      if (declString[i] !== ":") { continue }
+    // The extra characters tacked onto the end ensure that there is a character to check
+    // after the colon. Otherwise, with `background:pink` the character after the
+    const propPlusColon = decl.toString().slice(0, endOfPropIndex) + "xxx"
+
+    for (let i = 0, l = propPlusColon.length; i < l; i++) {
+      if (propPlusColon[i] !== ":") { continue }
       locationChecker({
-        source: declString,
+        source: propPlusColon,
         index: i,
         lineCheckStr: decl.value,
         err: m => {